+++ src/mix/types.ts
@@ -17,8 +17,9 @@ import {BuildsParams, BuildsSearchParams} from './api/builds-types'
 import {DataHostsParams, DataHostsSearchParams} from './api/data-hosts-types'
 import {DeploymentFlowsParams, DeploymentFlowsSearchParams} from './api/deployment-flows-types'
 import {EnginePacksParams} from './api/engine-packs-types'
-import {EntitiesParams} from './api/entities-types'
+import {EntitiesParams, EntitiesSearchParams} from './api/entities-types'
 import {EnvironmentsParams, EnvironmentsSearchParams} from './api/environments-types'
+import {IntentsParams} from './api/intents-types'
 import {GeographiesParams, GeographiesSearchParams} from './api/geographies-types'
 import {JobsParams, JobsSearchParams} from './api/jobs-types'
 import {LanguageTopicsParams} from './api/language-topics-types'
@@ -38,6 +39,7 @@ export * from './api/engine-packs-types'
 export * from './api/entities-types'
 export * from './api/environments-types'
 export * from './api/geographies-types'
+export * from './api/intents-types'
 export * from './api/jobs-types'
 export * from './api/language-topics-types'
 export * from './api/literals-types'
@@ -59,6 +61,7 @@ export type MixRequestParams =
   | EntitiesParams
   | EnvironmentsParams
   | GeographiesParams
+  | IntentsParams
   | JobsParams
   | LanguageTopicsParams
   | LiteralsParams
@@ -75,6 +78,7 @@ export type MixRequestSearchParams =
   | BuildsSearchParams
   | DataHostsSearchParams
   | DeploymentFlowsSearchParams
+  | EntitiesSearchParams
   | EnvironmentsSearchParams
   | GeographiesSearchParams
   | JobsSearchParams